fix: resolve all clippy warnings and improve code quality
- Fixed 'parameter is only used in recursion' warnings by making methods static
- Replaced match expressions with matches\! macro where appropriate
- Fixed naming conventions (LSM -> Lsm)
- Added Default implementations for WriteBatch and WaitFreeReadBuffer
- Renamed iterator methods from next() to advance() to avoid confusion
- Created type aliases for complex types
- Fixed loop variable indexing issues using iterators
- Removed empty lines after doc comments in binaries
- Used struct initialization pattern instead of field assignments
- Fixed various other clippy suggestions for idiomatic Rust

All library and binary code now compiles without clippy warnings.

+++ src/transaction/version_cleanup.rs
@@ -144,7 +144,7 @@ mod tests {
                 format!("key_{}", i).into_bytes(),
                 Some(format!("value_{}", i).into_bytes()),
                 now + i,
-                i as u64,
+                i,
             );
         }
